Refactor the time series classes
They now inherit from a single TimeSeries class. The fixed and variable
index middleman classes are now gone since they were not actually used
anywhere and most functionality was overriden in the time series
classes anyway. This made it easier to deal with the fact that
'ignore_year' and 'repeat' are actually applicable to the variable step
time series as well.

Re #56