Fix bug when updating objects
When updating an object only existing in the original tables only the data from the "Entity" table was copied to the "DiffEntity" table,
however when querying for the changed object in the object_sq it was filtered by the "DiffObject" table wich nothing was copied to.
Fixed by changing the filter to use the "entity_type" column instead.