Re-fix bug when updating objects
Supersed fix in Commit 7589007b, which works, but actually hides a 
mistake in other parts of the code.

The current fix consists in not marking as dirty and updated records 
that weren't updated. 

For instance, when updating objects, we do not touch the object table at 
all --just the entity table. So in this commit, we don't mark any 
records as dirty in the object table. As result, the `object_sq` 
property is still able to get records from the original (non-diff) 
untouched table.