Fix failing tests for DiffDatabaseMapping, rasing "Nothing to commit" error
Test where failing because the test setUp method was reseting the mapping and calling commit_session
after each test. However the commit_session raises an error "Nothing to commit" when commiting where
there are no changes. The first test setUp it runs will not have any changes and error is raised.

Maybe a more robust way would be to create a new database for each test instead of sharing the
test database.