Speed up datetime parsing when reading database json timeseries.
When converting a database timeseries json to TimeSeriesVariableResolution class
the datetime parsing took a very long time. By not using dateutil parser and instead
relying on the numpy.datetime64 parser there was a ~50 times speedup.

This means that the datetime strings in a valid database json must be iso8604 format
otherwise the numpy.datetime64 might fail.