Deal with immutable fields more nicely
Re #69

Now it is Ok to update items with immutable fields as long as the
updated value is the same as the current value. This has the effect
that we don't need to pop eg the 'class_id' entry from the dictionary
item when updating an object. The new requirement is that the 'class_id'
needs to be correct (in the end, we never want to change an object's
class, or at least not for now).