Implement options for parameter mapping
Certain parameter value types, like time series, have some additional
options available. When mapping the values from a tabular source it
should be possible to construct the parameter values with the given
options. As a humble start, this commit introduces the minimum needed
facilities for this allowing a client to set the 'repeat' flag for time
series type data. This is in response to user request, see issue
toolbox/#521