A few improvements to the update mixin
We no longer modify the arguments passed to the function. This is
important so the client can keep using the arguments for something
else.

Also, we now check that there's effectively something to update
before proceeding. This is important so the client knows exactly
what was updated.

Finally, we remove an obsolete method and bump the version.