Add support for composite primary keys in DiffDatabaseMapping
We now allow for marking items as dirty according to a 'composite'
primary key, that is, a primary key composed of multiple fields.
For now this is only used in updating one of the member objects of
a relationship, where we mark a specific combination of entity_id plus
dimension as dirty. May be overkill, but it also may help us with
the problem of updating value lists. Again, nothing too impactful
especially since we may consider getting rid of the diff approach
anyways.