More flexible column header mapping
json_mapping allowed column headers be referenced by their names or
indexes. The distinction was made according to whether
ColumnHeaderMapping.reference was an int or a str. However, the
reference set by Toolbox was always str so it was impossible to map a
column header by index. We now try to convert the string to an integral
index if the string is not found among the headers enabling one to give
the index as a string not only as an int.