Introduce the transaction ids to the ids_for_in table
It may happen that the `DatabaseMapping.in` method is used multiple
times within a single query. When this happens, it looks like the
session has no time to clear the `ids_for_in` table in between uses,
so we need to isolate these uses otherwise. This commit introduces
an additional id to isolate each use of the table. Doesn't feel
awesome but seems to do the job for now.