Initial support for data-level convenience groups
Re #94

Similar to what we were doing in SpineOpt.jl, we also use an `entity`
here to represent the group. Membership is established by populating
the `group_entity` table with tuples `entity_id`, `entity_class_id`,
`member_id`.
Then, for example, a unit group is just a unit entity, which appears
under the `entity_id` column of the `group_entity` table.

This has the absolutely huge advantage that we don't need to
port definitions of parameters and relationship classes
from the entity to the group. The group is already an entity,
thus it has access to all relationship classes and parameters
that it needs.