Make commit_id columns non-nullable
When creating or upgrading a database we now add a commit corresponding
to the change. This allows us to refer to this commit in the base
alternative as well as the initial entity types and entity class types.
We don't need to set the commit_id columns to nullable in alternatives,
entity_type and entity_class_type tables anymore