Finalize support for groups import/export and introduce Anyone
The purpose of Anyone is being able to omit the `IN` clause, by passing
`Anyone` in the set of elements to match. Any element matches Anyone
so the `IN` clause can be omitted. We were using `None` for this
purpose before and that seemed to obscure. Let's see.

Bump version patch.