Finalize support for ObjectGroupMapping
Re #98, toolbox#751

We extend dict_to_map so it's able to return ObjectGroupMappings.
We also introduce the property `import_objects` to ObjectGroupMapping
which works in the same way as for RelationshipClassMapping.

Finally, we introduce `EntityClassMapping.from_instance` which
basically allows for converting between instances of
`EntityClassMapping` subclasses.

Bump version patch.