Simplify import scenario alternatives
Instead of the caller having to manage the rank explicitly, now they
can provide the name of an alternative which marks the 'insertion'
position. So all the alternatives are inserted before that one.

I've seen this approach for inserting QActions to a QMenu
and I think it works quite well.