Single scenario only in alternative_value_filter
alternative_value_filter was picking all alternatives of all active
scenarios. This turned out to be incorrect operation; clients are
supposed to choose a single scenario to work on.
apply_alternative_value_filter() now requires one to give a single
scenario, alternatives of which are used for selecting values or
overriding alternatives.

Re #100