Fix bug with scenario/alternative migrations script
Re toolbox#778

For some reason, there are dbs without the expected unique constraint
in the parameter value table that enter the migration process. The
refereed issue reports the error and traceback.

To fix this, we manually check that the constraint exists before
removing it, which is not awesome but the only way to cut loses here
at the moment.