Introduce QuickDatabaseMapping
This new class can be used as a replacement for DiffDatabaseMapping
or even DataBaseMapping (although it doesn't ship the ...QueryMixin).

The idea is to speed up construction by not generating the ORM
instances. Queries are performed directly over the Table objects
in the _metadata attribute. Modifications to the db are done in a
single transaction.

Clients that need a bounded db handle to write some stuff
(Importer, Combiner) can use this class to see
some speedup.