Support for in-memory databases
The biggest issue with in-memory sqlite databases is that you cannot
refer to a specific database by its URL. It is now possible to pass an
engine to DatabaseMappingBase which allows in-memory database. Filters
don't work reliably, if multiple database mappings share the same
engine, though.

Unit tests have been modified to use in-memory databases where
applicable. The tests now take around 20% less time to execute on my
system.

Re #114