Drop engine parameter from DatabaseMappingBase
We can create in-memory databases just by giving 'sqlite://' as the url
and set create=True. The use cases for shared engines between database
mappings are not very common (there are a few in unit tests, though) and
it is possible that url does not correspond to the engine argument which
may lead to inconsistencies.

This drops the 'engine' parameter in favor of url='sqlite://` and
create=True.

Re #114