Add support for filtering 'write' operations by a given scenario
Re #112

DatabaseMappingBase subclasses now provide a method to get the
default alternative for parameter value import operations. This method
is used of course by the corresponding import_functions, but it
can also be overriden by the scenario filter. In this way we achieve
the goal of filtering 'write' operations, not only 'read' operations.