Create filter scenario if it doesn't exist
This is to support creating an array of ProjectItemResources
for a specific scenario. Sometimes resources come from different
places and they don't have all the same active scenarios, but
still we want to create a consistent set of resources for each such
scenarios. This commit supports that by storing the scenario name
in the filter 'state' even if it doesn't exist. Then, when the
user finally decides to write to a scenario filtered db, the
scenario is created so it all works out as the user probably expects.