Fix synch problem when applying the scenario filter
Because of how things work at the moment with the filtering of
'predecessor' and 'successors' resources in toolbox and engine,
the application may apply the scenario filter to a successor resource
and *then* write that scenario to the db.

For this reason we need to re-check if the scenario is there
when creating the default import alternative.

The fact that we need this may indicate a design flaw,
will try to think about it.