Introduce execution filter
Re #112

The scenario filter no longer overrides the import alternative.
Instead, we introduce the execution filter to more explicitly
override the import alternative in the context of an execution,
that may involve *multiple scenarios from different resources*.

The import alternative becomes tied to all the scenarios from
the corresponding execution so the user can keep track.
Maybe not perfect yet, but a small improvement compared to what
we had before I believe.