Add initial support for MySQL (experimental)
Re #120

1. Fix problems in schema (missing indexes, too long identifier names,
not long enough text column types, invalid referential actions, etc.)
2. MySQL can't use the same temp table more than once in a SELECT.
This affects queries with our diff tables which need to be temporary
so they're *always* deleted at the end of the session (even if the
program crashes). Making the diff tables permanent and drop them
ourselves when closing the connection is risky, because the program
may crash before we get to that point; which would require then
to have a system to clean up orphan diff tables (not easy).
Instead, we keep the diff tables as temp and use a compound select
as replacement whenever we need to SELECT on them.
3. Introduce `group_concat()` which allows to use ORDER BY in the
case of MySQL.
4. MySQL doesn't suffer from the 'too many sql variables' as does
SQLite, so we adapt `DatabaseMapping.in_()` accordingly.