Introduce export_functions.export_expanded_parameter_values()
Re toolbox#783

The idea here is to facilitate the excel export in
spine_engine.spine_io. But also, I think this can be used for other
purposes such as the 'query' functionality that has been discussed.
For this reason, it's being implemented here.

Another side change is the renaming of `Anyone` to `Asterisk`, trying
to make it more understandable and usable.