Introduce filter_re as a property of export Mapping
Re Spine-project/Spine-Toolbox#1260

The idea here is to filter the output of the mapping (what comes
out of _query()), by a given regular expression.

More 'simple' filtering can be build on top of this, I mean, if
we don't want to use regular expressions and prefer to use a
simpler syntax of our own (like comma separated list of
accepted values), we could translate that into a regexp for
performance.