Add support to limit the number of tables produced by a writer
Re Spine-project/Spine-Toolbox#1267

Sometimes, when the user selects Position.table_name for a mapping
that outputs a very large amount of db_rows, the Exporter spec
editor takes a long time in loading the preview. Before, it was
possible to limit the number of rows that go into a table but not
the amount of tables itself.