Ignore empty strings as parameter values when importing
This is a tough one. The old importer API did it this way. I'm
not sure what's best here but let's go with this for now, to
help compatibility. I believe ideally we should be able to import
empty strings. Maybe what we need is a system so the user can
define what's a missing value...