Introduce ParameterValueTypeMapping for exporter
This mapping yields the type, somehow consistently with importer.
So the options are, time_series, time_pattern, array, xd_map (where x
is the dimension count), and then single_value for all the rest.

If sent to the table_name position, it also filters all data in
that table by the corresponding type. This is all a bit arbitrary
to satisfy the needs of the good old export to excel.