Reword object group export mappings
It was not possible to export object groups alone, i.e. you got one
group row for each object in that group regardless of whether the
objects were hidden or not. We now use specialized queries for object
groups and member objects so that they work like entity classes and
entities.

Re Spine-project/Spine-Toolbox#1332