Fix pivot export padding, header buddies
In the special case when there was a pivot header but no 'left side' in
the pivot table, a pivot table was exported with the pivoted data
starting from the first column which should have been empty because of
the header. We now pad the pivot table accordingly.

Also, there was an issue with header buddies in cases where there were
two or more different types of child mappings applicable as buddies. For
example, ObjectClassMapping could equally well be a header for objects
or parameter definitions. We now always offer only a single buddy
option.

Re Spine-project/Spine-Toolbox#1340