Fix filtering with hidden tail
Filtering by regular expression didn't work as expected for hidden tail
mappings because we used to drop hidden tail mappings unconditionally.
We now leave the tail as is if filter_re is set; the change required
modifications to pivot tables as well as now it is possible that the
hidden tail writes data to the pre-pivot regular table.

Re Spine-project/Spine-Toolbox#1404