Rewrite value transformation query
Parameter value transformation manipulator used to append transformation
instructions to the value JSON. Applying the instructions required
calling from_database() making the manipulator hard to use for client
code that doesn't have access to from_database() or wants to deal with
raw JSON instead.

We now calculate the transformation when the manipulator is applied to
the database and save them as JSON in manipulator state. In the subquery
we directly replace parameter's value and type by the precalculated
values removing the from_database() magic.

Re #132