Reduce the amount of querying needed for several db operations
Querying large dbs is quite expensive. We attempt to reduce
querying by introducing a number of measures.

First, we add the option to provide a 'db-cache' to methods that
do integrity checks, so they can use this cache as a db replacement.
This is mainly intended to be used with the cache from SpineDBManager
(toolbox)

Second, we add the option to return items rather than ids in methods
that add and update items. This is so the client doesn't need to
re-query the db to get changes.

Third, methods that update items now expect to receive the full item.
Before, only the changed fields were necessary, and we would obtain
the rest by querying the db. So now, it is up to the client to
figure out all the fields, but it should be easy for them since they
must have queried the db before attempting any update.