Simplify filtering by title state in Exporter
Re Spine-project/Spine-Toolbox#1527

I was a bit worried about this since we introduced it: Basically,
to be able to filter the ExportMapping query by title state,
we were creating a subquery and then querying it again.

Coupled with the query overriding from the scenario and tool filter,
this led to parser stack overflow.

Now we use our custom _FilteredQuery class to bypass this.