Fix GdxWriter.finish_table() in failure case
GdxWriter.start_table() can fail if one tries to append to an existing
table. This used to leave GdxWriter in an unspecified state which could
lead to a Traceback in GdxWriter.finish_table(). finish_table() in turn
is called write() even if writing fails to make sure the table is
properly closed creating a hard crash in GdxWriter.

We now avoid any work in GdxWriter.finish_table() if start_table()
failed to make sure we don't try to finish an already broken table.

Re Spine-project/Spine-Toolbox#1483