Fix exporting empty GAMS parameters
GdxWriter identified GAMS parameters by checking that the bottom right
element in output table was a number. The test failed, or was actually
skipped if the table was empty resulting in parameters being written as
sets. We now use the table header to identify parameters as the last
column header is expected to be left empty in this case.

Re #144