Fix empty tables in export mappings
Export mappings that query parameter values don't return any database
rows if there are no values. This prohibited export mappings from
generating empty tables or tables with just header rows. To circumvent
the problem, we now drop the tails of mapping hierarchies that are not
needed for table names or column headers when creating the corresponding
queries.

Re #150