Extend column conversion for wide pivots
get_mapped_data() has a column_convert_fns argument that contains
conversion functions for columns. It is not an uncommon case, though,
to try to import pivoted data that has more columns than specified for
the convert functions, e.g. large time series resulting in the last
column not getting any conversion at all. As a fallback, we now
pick the last specified conversion function and use that as the default
conversion for columns that don't have a specified conversion.

Re Spine-project/Spine-Toolbox#1654