Don't unhide RelationshipMappings when pivoting
When exporting indexed relationship parameter values, relationship names
are usually hidden. However, this caused in certain cases the values to
end up in separate rows in the output table. This was caused by pivoting
being conservative and handling these values as if they had different
indices. We now refrain from giving hidden RelationshipMapping a column
when the regular table is created for pivoting.

Re Spine-project/Spine-Toolbox#1687