Use unparse_value to disambiguate method format for import
The idea here is to use the unparse_value argument to bring the method
into DB format before calling from_database on it. unparse_value should
be consistent with the format the data is in. So if the method is given
as a Python object (the default I guess), then unparse_value should be
to_database (the default). But if the method is given as a list of
value and type (the DB server way), then unparse_value should be
spine_db_server._unparse_value.

I know this is a bit messy but it's the current design and the outcome
of several patches (one day we should consider rebuilding things more).