Reject value list changes if parameter values exist
Value list is not a fixed field for parameter definitions anymore.
This makes it possible to update the value list, though we reject updates
that change the value list when the parameter has values. Allowing
one to change the value list while values are present would force us
to validate and update all values, too. We might allow that in the
future, but for now we just prevent the case where values are not
from the list.

Re Spine-project/Spine-Toolbox#1793