Introduce id generator and order entities in import mapping so it works
Since now we import objects and relationships together via the
"entities" key, we need to make sure the elements are in the right order
so 0-dimensional entities are processed before multidimensional ones
(because the latter may need the former). This implies even assigning
an id to the 0-dimensional ones *before* they are actually added to
the db, and this is done with a context manager.