Simplify TempId resolution
Re Spine-Toolbox/issues#2295

TempId are no longer replaced when resolved. Instead, we keep
a mapping from TempId to db-id that we use whenever we interact
with the db, and only then. The TempId stays in place on the cache-side
for as long as the cache lives.

This removes the need for TempIdDict.