Simplify refresh
Spine-Toolbox#2300

Instead of clearing the cache and persisting the items,
we just keep the cache intact and then check, as items are fetched
from the DB, whether those items are already in the cache and in that
case skip them.

These guarantees that items never lose their TempID. Maybe we should
rename it to CacheId.