Fix db requiring committing twice
The need for double committing arose from the fact that if an
entity or entity class was re-added after it had already been
deleted from the db, some of its data was still present in
the db. That caused some unique constraint errors.

- Previously when an entity was deleted, only the entity table
was considered. Now the data corresponding to the entity is
also deleted from the entity_element -table.

- When deleting an entity class, the data form the
entity_class_dimension -table corresponding to the classes id
is also deleted.

Re spine-tools#Spine-Toolbox#2296