Invalidate cache item id instead of deleting it
We shouldn't delete cache item's "id" key when rolling back changes
because clients expect the id to be present. This was an issue in
Toolbox where the Database editor's views would handle removed items
expecting to have the "id" key available after it was deleted in
spinedb_api.

We don't touch the id anymore but use a new 'is_id_valid' flag to
signal that the id has gone out of scope.

Re spine-tools/Spine-Toolbox#2291