Implement rudimentary "merge conflict resolution"
We now apply a rudimentary conflict resolution when fetching items
that already exist in the db cache. Conflict resolution is handled
by a callback that could let user decide what to do in the future
but is currently used to force the default behavior, which is to
keep cache as-is.

Re spine-tools/Spine-Toolbox#2431