Use in-memory ids instead of database ids
This work fixes a lot of issues where the database has been modified externally
between fetches/commits.

The in-memory cache of DatabaseMapping now uses its own ids for items called
item ids, i.e. every item now has a temporary id. However,
instead of a specific TempId type we use negative integers for
simplicity.

There are mappings from item ids to database ids and back. Every time
an item is added to the mapping tables, a new item id is created. When
fetching or committing, we map the database id to the item id. We also
check more rigorously whether a fetched item is actually the same item that
may already be in the mapped tables.

Re spine-tools/Spine-Toolbox#2431