Fix opening infinite db editor tabs from same db
If a db_map is already opened in a db editor, new tabs
with the same db_map won't be opened anymore. Instead
the one already open is brought up. If however the
same db_map is opened in a new window, it is allowed
to open.

Re spine-tools/Spine-Toolbox#2435