Rewrite the pages about metadata in the docs
This replaces "Metadata description" and "Result metadata description"
pages in the documentation by a single "Metadata" paget.
The new page contains a short introduction to metadata
and some suggestions on what kind of stuff to store there.
This is in contrast to the datapackage-like JSON definition that the
two removed pages contained.

Re #329