Don't show removed stuff after a refresh
When the user refreshes session and there are external changes,
we mark all committed items as compromised - since we don't know
if they are still in the DB.
Then as we fetch items, we mark them as uncompromised (that is,
back to committed).

Compromised items are considered invalid so unless they are refetched,
they stay in the dark.

This doesn't apply to uncommitted items since these are
considered for the future, so they shouldn't be affected by newer
state of the DB. Let's see.