Fix parameter type export mappings
ParameterValueTypeMapping and ParameterDefaultValueTypeMapping are
supposed to return true types now that the information is available
in databases for scalar types as well. However, many databases do
not yet have this information as we do not enforce it, yet.
In these cases, the mappings need to parse the value manually so
they can return some data for the row.