Fix marking added item 'to_update'
In a previous commit, a check if item's mapped table was purged
was replaced by checking if it's corresponding existing (but removed)
item wasn't committed. Turns out there are cases where we still need
to check if the mapped table isn't purged before we can mark
the added item as 'to_update'.