Fix unique (now required) key validation
It was possible to add items without actually providing values for
keys that were required to commit the item.
One such case was scenario alternatives which could be added
without specifying 'rank'.

Now the items have a required_key_combinations field that is used
to check for missing keys. It replaces the unique_fields and
coresponding_unique_id_keys fields in this role and allow much
more flexible definition of which keys and key combinations
are required when adding new items.