Fix replacing a committed but then removed item with a new item
Previously, if a new item would replace a removed item (fetched from
database, then removed), we would just mark the new item as 'to_update'.
However, if the new item was later removed,
commit would fail because the new item never didn't have a DB id.

Now, instead of 'stealing' the removed item's identity by setting
a new item's status to 'to_update', we keep the status as 'to_add'
and store the removed item for later removal on commit.
This way the new items replaces the removed on in a clean way.