Add get/add/update dataframe interoperability functions
Added functions to get/add/update mapped items directly.
These functions work on a "higher" level than fetch_as_dataframe()
which works with database queries and is read-only.
Entity class name is also now included in the dataframe so
we can identify the entity uniquely which is not possible
with byname only.

Re #353