Turn off field type sniffing in DatapackageReader
Field type sniffing in frictionless is not 100% reliable since
it sniffs only a limited number of rows. This will create problems
if the sniffer declares a field as integers but in fact it contains
floats.
In any case, we will cast the values to the desired type later in
the importer pipeline.

Re spine-tools/Spine-Toolbox#3154