Fix alternatives filter to handle entity_alternative more thoroughly
We had two issues:
- If an entity had an entry in entity_alternative and we filtered
by another alternative, the entity was dropped by the filter regardless!
- We didn't handle the case where an entity is active in, say,
alternative A, and inactive in alternative B, and the alternatives
filter list includes both alternatives. Here we solve it by assuming
the alternatives in the list are sorted by increasing rank and apply
the same logic as the scenario filter.