Handle broken references in DB
Looks like historically we have allowed our DBs to hold
broken references - basically a failure to cascade remove
in any of our multiple API versions.

The current API raises a RuntimeError when this occurs but to me
it seems so common, that we should better handle it.
And the way I am handling it here is to just skip the record
with the broken reference. The implementation at the moment
goes a bit in circles by adding the item and then removing it
if a broken ref is found. This is because to be able to check references
at the moment we need to add the item.