Fix add items from DB and refresh
Fetching an item that was equivalent to one from the memory mapping
created a duplicate instead of merging. Then, committing resulted
in unique constraint violations.
Seems like we carried this one from long ago.

Also, refreshing didn't remove all the unique key values
due to the order in which we do things.