Simplify code in spine_db_server
The _DBManager class was creating extra noise which made it more
difficult to debug. We can also store the _DBWorker directly in the
DBServer class which avoids us keeping a dict of workers.
We also get rid of _ManagerRequestHandler and replace it by top-level
functions which seems to make more sense.