Fix broken entity references
We have DBs with entities where the elements are some other entities
no longer in the DB. This must be due to an error with cascade removing
in a previous version of the API, so we handle it here.
We just check that the element id count is equal to the element name
count.