Fix conflicts correctly when committing to a filtered DB
In such case we need to fetch from the unfiltered DB to avoid
unique constraint errors.

For example, say we have filtered by an alternative resulting
in an entity being filtered out from entity subquery.
Then the mapping would allow us to add the same entity again
but because it is filtered it will never find out that is a duplicate,
then at the moment of committing we find the unique constraint error.